pyproject.toml
[tool.poetry]
name = "WindWith2021"
version = "1.4.1"
description = "简书 2021 年终总结"
authors = ["FHU-yezi <yehaowei20060411@qq.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
jieba = "^0.42.1"
JianshuResearchTools = "2.10.0"
pywebio = "^1.6.0"
pandas = "^1.4.2"
peewee = "^3.14.10"
PyYAML = "^6.0"
openpyxl = "^3.0.10"
pyecharts = "^1.9.1"
wordcloud = "^1.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
